# backend/openapi.json
{
  "components": {
    "schemas": {
      "AIAdvisorResponse": {
        "description": "Response from POST /advisor/ask",
        "properties": {
          "category": {
            "default": "general",
            "title": "Category",
            "type": "string"
          },
          "context_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "fallback_used": {
            "default": false,
            "title": "Fallback Used",
            "type": "boolean"
          },
          "follow_ups": {
            "items": {
              "type": "string"
            },
            "title": "Follow Ups",
            "type": "array"
          },
          "insights": {
            "items": {
              "type": "string"
            },
            "title": "Insights",
            "type": "array"
          },
          "latency_ms": {
            "default": 0,
            "title": "Latency Ms",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "provider": {
            "default": "fallback",
            "title": "Provider",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "query",
          "summary",
          "insights",
          "recommendations",
          "follow_ups"
        ],
        "title": "AIAdvisorResponse",
        "type": "object"
      },
...
      "AdvisorQueryRequest": {
        "description": "Body for POST /advisor/ask",
        "properties": {
          "conversation_history": {
            "description": "Prior turns in this conversation, oldest first. Sent to the LLM so it can maintain context across messages. Limit to last 6 turns to avoid token overflow.",
            "items": {
              "$ref": "#/components/schemas/ConversationTurn"
            },
            "title": "Conversation History",
            "type": "array"
          },
          "include_optimization": {
            "default": false,
            "description": "Include optimization-aware narrative in context",
            "title": "Include Optimization",
            "type": "boolean"
          },
          "include_snapshots": {
            "default": true,
            "description": "Include snapshot history in context",
            "title": "Include Snapshots",
            "type": "boolean"
          },
          "portfolio_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Portfolio to analyse (uses active if omitted)",
            "title": "Portfolio Id"
          },
          "query": {
            "description": "User's natural language question",
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "AdvisorQueryRequest",
        "type": "object"
      },
...
      "ContextSummary": {
        "properties": {
          "has_recent_changes": {
            "title": "Has Recent Changes",
            "type": "boolean"
          },
          "holdings_count": {
            "title": "Holdings Count",
            "type": "integer"
          },
          "sectors_count": {
            "title": "Sectors Count",
            "type": "integer"
          },
          "snapshots_count": {
            "title": "Snapshots Count",
            "type": "integer"
          }
        },
        "required": [
          "holdings_count",
          "snapshots_count",
          "sectors_count",
          "has_recent_changes"
        ],
        "title": "ContextSummary",
        "type": "object"
      },
      "ConversationTurn": {
        "description": "A single turn in the conversation history for multi-turn AI calls.",
        "properties": {
          "content": {
            "description": "Message text",
            "title": "Content",
            "type": "string"
          },
          "role": {
            "description": "'user' or 'assistant'",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ConversationTurn",
        "type": "object"
      },
...
      "FinancialRatiosResponse": {
        "description": "Bundled fundamentals response \u2014 per-holding ratios + weighted portfolio metrics\n+ trust metadata + backend-owned thresholds.\n\nHoldings with unavailable fundamentals are included with source='unavailable'\nand an error field rather than being silently dropped.\n\n`thresholds` ships the canonical threshold constants from\nfundamentals_view_service.py to the frontend. The frontend uses these\nfor traffic-light coloring and insight rules and never hardcodes them.",
        "properties": {
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/FinancialRatioResponse"
            },
            "title": "Holdings",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/FundamentalsMeta"
          },
          "thresholds": {
            "$ref": "#/components/schemas/FundamentalsThresholds"
          },
          "weighted": {
            "$ref": "#/components/schemas/WeightedFundamentals"
          }
        },
        "required": [
          "holdings",
          "weighted",
          "meta",
          "thresholds"
        ],
        "title": "FinancialRatiosResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PortfolioContextPayload": {
        "description": "Clean, LLM-friendly context object.\nReturned by GET /advisor/context/{portfolio_id} for debug visibility.\nNever exposes raw ORM objects.",
        "properties": {
          "built_at": {
            "title": "Built At",
            "type": "string"
          },
          "diversification_score": {
            "title": "Diversification Score",
            "type": "number"
          },
          "hhi": {
            "title": "Hhi",
            "type": "number"
          },
          "max_holding_ticker": {
            "title": "Max Holding Ticker",
            "type": "string"
          },
          "max_holding_weight": {
            "title": "Max Holding Weight",
            "type": "number"
          },
          "num_holdings": {
            "title": "Num Holdings",
            "type": "integer"
          },
          "num_sectors": {
            "title": "Num Sectors",
            "type": "integer"
          },
          "portfolio_id": {
            "title": "Portfolio Id",
            "type": "integer"
          },
          "portfolio_name": {
            "title": "Portfolio Name",
            "type": "string"
          },
          "recent_changes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecentChanges"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_profile": {
            "title": "Risk Profile",
            "type": "string"
          },
          "sector_allocation": {
            "items": {
              "$ref": "#/components/schemas/SectorBrief"
            },
            "title": "Sector Allocation",
            "type": "array"
          },
          "snapshot_count": {
            "title": "Snapshot Count",
            "type": "integer"
          },
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/SnapshotBrief"
            },
            "title": "Snapshots",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceMetaPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "top3_weight": {
            "title": "Top3 Weight",
            "type": "number"
          },
          "top_holdings": {
            "items": {
              "$ref": "#/components/schemas/HoldingBrief"
            },
            "title": "Top Holdings",
            "type": "array"
          },
          "total_cost": {
            "title": "Total Cost",
            "type": "number"
          },
          "total_pnl": {
            "title": "Total Pnl",
            "type": "number"
          },
          "total_pnl_pct": {
            "title": "Total Pnl Pct",
            "type": "number"
          },
          "total_value": {
            "title": "Total Value",
            "type": "number"
          }
        },
        "required": [
          "portfolio_id",
          "portfolio_name",
          "source",
          "total_value",
          "total_cost",
          "total_pnl",
          "total_pnl_pct",
          "num_holdings",
          "top_holdings",
          "sector_allocation",
          "risk_profile",
          "hhi",
          "diversification_score",
          "max_holding_ticker",
          "max_holding_weight",
          "top3_weight",
          "num_sectors",
          "snapshot_count",
          "snapshots",
          "recent_changes",
          "built_at"
        ],
        "title": "PortfolioContextPayload",
        "type": "object"
      },
...
      "RecentChanges": {
        "properties": {
          "added_tickers": {
            "items": {
              "type": "string"
            },
            "title": "Added Tickers",
            "type": "array"
          },
          "days_apart": {
            "title": "Days Apart",
            "type": "integer"
          },
          "decreased_count": {
            "title": "Decreased Count",
            "type": "integer"
          },
          "increased_count": {
            "title": "Increased Count",
            "type": "integer"
          },
          "removed_tickers": {
            "items": {
              "type": "string"
            },
            "title": "Removed Tickers",
            "type": "array"
          },
          "value_delta": {
            "title": "Value Delta",
            "type": "number"
          },
          "value_delta_pct": {
            "title": "Value Delta Pct",
            "type": "number"
          }
        },
        "required": [
          "days_apart",
          "value_delta",
          "value_delta_pct",
          "added_tickers",
          "removed_tickers",
          "increased_count",
          "decreased_count"
        ],
        "title": "RecentChanges",
        "type": "object"
      },
...
      "RiskMetrics": {
        "description": "Scaffold for risk analytics response.\nPopulated by analytics/risk.py in Phase 2.",
        "properties": {
          "beta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beta"
          },
          "max_drawdown": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Drawdown"
          },
          "note": {
            "default": "Risk metrics require historical price data. Available in Phase 2.",
            "title": "Note",
            "type": "string"
          },
          "sharpe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe Ratio"
          },
          "var_95": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Var 95"
          },
          "volatility_annualised": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volatility Annualised"
          }
        },
        "title": "RiskMetrics",
        "type": "object"
      },
...
      "SectorBrief": {
        "properties": {
          "num_holdings": {
            "title": "Num Holdings",
            "type": "integer"
          },
          "sector": {
            "title": "Sector",
            "type": "string"
          },
          "weight_pct": {
            "title": "Weight Pct",
            "type": "number"
          }
        },
        "required": [
          "sector",
          "weight_pct",
          "num_holdings"
        ],
        "title": "SectorBrief",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "P-Insight Portfolio Analytics API. Visit /docs for interactive API documentation.",
    "title": "P-Insight API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/advisor/ask": {
      "post": {
        "description": "Main entry point for AI-powered portfolio advisory.\n\n- Builds structured portfolio context from the database\n- Calls the configured LLM provider (Claude \u2192 OpenAI \u2192 fallback)\n- Returns summary + bullet insights + recommendations + follow-ups\n\nWhen fallback_used=True the frontend should run rule-based routeQuery()\nlocally and display its response instead.",
        "operationId": "ask_advisor_api_v1_advisor_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvisorQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAdvisorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask the AI Portfolio Copilot a question",
        "tags": [
          "AI Advisor"
        ]
      }
    },
    "/api/v1/advisor/context/{portfolio_id}": {
      "get": {
        "description": "Returns the clean JSON context object that the AI advisor uses.\nIntended for the /debug page so developers can inspect exactly\nwhat data is being provided to the LLM.",
        "operationId": "get_portfolio_context_api_v1_advisor_context__portfolio_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioContextPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview the portfolio context that would be sent to the AI",
        "tags": [
          "AI Advisor"
        ]
      }
    },
...
    "/api/v1/analytics/commentary": {
      "get": {
        "description": "Return rule-based portfolio insights and commentary.\nPhase 1: Derived from summary and sector data.\nPhase 2: Fed as context to AI Chat module.",
        "operationId": "get_commentary_api_v1_analytics_commentary_get",
        "parameters": [
          {
            "description": "Data source mode",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "uploaded",
              "description": "Data source mode",
              "enum": [
                "uploaded",
                "live",
                "broker"
              ],
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get AI-style portfolio commentary",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/analytics/ratios": {
      "get": {
        "description": "Return fundamentals and valuation ratios for each holding, plus:\n  - portfolio-level weighted-average metrics (PE, PB, ROE, etc.)\n  - trust / completeness metadata (source, incomplete flag, unavailable tickers)\n\nHoldings with unavailable fundamentals are included in the response with\nsource='unavailable' and an error field \u2014 they are NOT silently dropped.\nThis prevents callers from presenting partial data as if it were complete.\n\nPer-holding fields:\n  Valuation: pe_ratio, forward_pe, pb_ratio, ev_ebitda, peg_ratio\n  Income:    dividend_yield\n  Quality:   roe, roa, operating_margin, profit_margin\n  Growth:    revenue_growth, earnings_growth\n  Balance:   debt_to_equity, market_cap\n  Trust:     source, error, fetched_at, cache_age_seconds\n\nPortfolio-level weighted fields:\n  wtd_pe, wtd_pb, wtd_roe, wtd_roa, wtd_ev_ebitda, wtd_peg,\n  wtd_div_yield, wtd_operating_margin, wtd_profit_margin,\n  wtd_revenue_growth, wtd_earnings_growth, wtd_debt_to_equity\n  + coverage counts per metric\n\nNull values for individual metrics indicate not applicable for that business\ntype (e.g. banks: null ev_ebitda, operating_margin, debt_to_equity).",
        "operationId": "get_financial_ratios_api_v1_analytics_ratios_get",
        "parameters": [
          {
            "description": "Data source mode",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "uploaded",
              "description": "Data source mode",
              "enum": [
                "uploaded",
                "live",
                "broker"
              ],
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialRatiosResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get financial ratios with portfolio-weighted metrics and trust metadata",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/analytics/risk": {
      "get": {
        "description": "Return portfolio risk metrics: beta, Sharpe ratio, volatility, drawdown.\nPhase 1: Returns scaffold response with note.\nPhase 2: Compute from historical price data.\nFull quant analytics (with live data) are available at /quant/full.",
        "operationId": "get_risk_metrics_api_v1_analytics_risk_get",
        "parameters": [
          {
            "description": "Data source mode",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "uploaded",
              "description": "Data source mode",
              "enum": [
                "uploaded",
                "live",
                "broker"
              ],
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskMetrics"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get risk metrics",
        "tags": [
          "Analytics"
        ]
      }
    },
...
  }
}